Merged in feature/filter_specific_af_after_hfr (pull request #836)
When filter wheel is available use only HFR trend for the current filter inside AutofocusAfterHFRIncreaseTrigger

* When filter wheel is available use only HFR trend for the current filter inside AutofocusAfterHFRIncreaseTrigger

* Filter AF after HFR trigger to exclude snapshots.

* adjust AUTs to check for filter specific hfr

* merge

* Show OriginalHFR | HFRTrend values additionally to percentages
Take the first image after Autofocus as "OriginalHFR" reference instead of the first one in the sample size to prevent slowly creeping up HFR trends to be ignored for Autofocus. When no autofocus is available, take the first history item instead.